requirements.txt
requests
flask
Flask-Discord
discord.py
"pymongo[srv]"==3.11